Add files via upload
This is the python logic for the mobile application. This file uses plyer to trigger notifications and haptic feedback when the user enters a geofence that has been classified as a high risk area. These geofences can be added through the radar.io website using a product key. This file also handles variables, altering the score based on user location changes.